fix methods for create, update and delete article

+++ src/main/java/com/it/academy/services/impl/ArticleServiceImpl.java
@@ -1,6 +1,8 @@
 package com.it.academy.services.impl;
 
 import com.it.academy.dao.ArticleDao;
+import com.it.academy.entities.Lesson;
+import com.it.academy.entities.User;
 import com.it.academy.exceptions.AppException;
 import com.it.academy.entities.Article;
 import com.it.academy.repositories.ArticleRepository;
@@ -8,7 +10,7 @@ import com.it.academy.services.ArticleService;
 import com.it.academy.services.LessonService;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -27,15 +29,15 @@ public class ArticleServiceImpl implements ArticleService {
     }
 
     @Override
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public List<Article> getAll() {
-        return articleRepository.findAll();
-    }
+    public Long create(Long userId, Article article, Long lessonId) {
+        Lesson lesson = lessonService.getById(lessonId);
+        User author = lesson.getSection().getCourse().getAuthor();
 
-    @Override
-    public Long save(Article article, Long lessonId) {
+        if(!userId.equals(author.getId())) {
+            throw new AccessDeniedException("You can't create article for this course!");}
+
         Article createdArticle = Article.builder()
-                .lesson(lessonService.getById(lessonId))
+                .lesson(lesson)
                 .title(article.getTitle())
                 .text(article.getText())
                 .build();
@@ -43,15 +45,24 @@ public class ArticleServiceImpl implements ArticleService {
     }
 
     @Override
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
-    public Long deleteById(Long id) {
+    public Long deleteById(Long userId, Long id) {
+        Article article = getById(id);
+        User author = article.getLesson().getSection().getCourse().getAuthor();
+
+        if(!userId.equals(author.getId())) {
+            throw new AccessDeniedException("You can't delete this article!");}
+
         articleRepository.deleteById(id);
         return id;
     }
 
     @Override
-    public Long update(Long id, Article updatedArticle) {
+    public Long update(Long userId, Long id, Article updatedArticle) {
         Article article = getById(id);
+        User author = article.getLesson().getSection().getCourse().getAuthor();
+
+        if(!userId.equals(author.getId())) {
+            throw new AccessDeniedException("You can't update this article!");}
 
         article.setText(updatedArticle.getText());
         article.setTitle(updatedArticle.getTitle());
